load_balancer/requirements.txt
fastapi
redis
uvicorn
google-cloud-storage
invoke
pymongo